+++ internal/config/root.go
@@ -62,10 +62,10 @@ func LoadFromDisk() Config {
 	ensureExists()
 	var config Config
 	configFile, err := os.Open(configPath())
-	defer configFile.Close()
 	if err != nil {
 		fmt.Println(err.Error())
 	}
+	defer configFile.Close()
 	jsonParser := json.NewDecoder(configFile)
 	jsonParser.Decode(&config)
 	return config